Add files via upload
Untitled11.ipynb is the code of training VGG16 teacher network and MobileNet student network. teacherres.ipynb is the code of training ResNet 18 teacher network and KD from ResNet18 to MobileNet kd_vgg.ipynb is KD from VGG16 to MobileNet kd_aver.ipynb is the equal average of the output of two teachers.